+++ tests/test_simple.py
@@ -5,6 +5,7 @@ Test simple APIs
 # %% Imports and settings
 import starsim as ss
 import sciris as sc
+import numpy as np
 
 n_agents = 1_000
 
@@ -126,10 +127,10 @@ def test_components():
 def test_parallel():
     """ Test running two identical sims in parallel """
     pars = make_sim_pars()
-    s1 = ss.Sim(pars)
-    s2 = ss.Sim(pars)
-    # s1, s2 = ss.parallel(s1, s2).sims
-    # assert np.allclose(s1.summary[:], s2.summary[:], rtol=0, atol=0, equal_nan=True)
+    sims = ss.MultiSim([ss.Sim(pars, label='Sim1'), ss.Sim(pars, label='Sim2')])
+    sims.run(keep_people=True)
+    s1, s2 = sims.sims
+    assert np.allclose(s1.summary[:], s2.summary[:], rtol=0, atol=0, equal_nan=True)
     return s1, s2