+++ CHANGELOG.rst
@@ -14,7 +14,8 @@ Version 0.2.3 (2024-02-26)
 - Removes ``STI`` class
 - Changes default death rate from units of per person to per thousand people
 - Allows ``ss.Sim(demographics=True)`` to enable births and deaths
-- *GitHub info*: PR `308 <https://github.com/amath-idm/stisim/pull/308>`_
+- Fix pickling of ``State`` objects
+- *GitHub info*: PRs `305 <https://github.com/amath-idm/stisim/pull/305>`_, `308 <https://github.com/amath-idm/stisim/pull/308>`_
 
 
 Version 0.2.2 (2024-02-26)